requirements.txt
grpcio
grpcio-tools
protobuf==5.27.2  # Adjust this based on compatibility
